(issue14) Move buddy to respective app/agents folder
